pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "webull-openapi-mcp"
version = "0.1.0"
description = "MCP Server for Webull OpenAPI - enables AI assistants to securely access Webull trading and market data"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
keywords = ["webull", "mcp", "openapi", "trading", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]
dependencies = [
    "fastmcp==3.0.2",
    "click",
    "python-dotenv",
    "webull-openapi-python-sdk>=2.0.2",
]

[project.scripts]
webull-openapi-mcp = "webull_openapi_mcp.cli:main"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "hypothesis",
    "pytest-mock",
]

[tool.hatch.build.targets.wheel]
packages = ["webull_openapi_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::DeprecationWarning",
]
